+++ frontend/src/pages/RecipientRecord/pages/CommunicationLog.js
@@ -19,7 +19,6 @@ import {
   methodFilter,
   resultFilter,
 } from '../../../components/filter/communicationLogFilters';
-import SomethingWentWrongContext from '../../../SomethingWentWrongContext';
 
 const COMMUNICATION_LOG_PER_PAGE = 10;
 const FILTER_KEY = 'communication-log-filters';
@@ -44,7 +43,6 @@ export default function CommunicationLog({ regionId, recipientId }) {
   });
 
   const { user } = useContext(UserContext);
-  const { setErrorResponseCode } = useContext(SomethingWentWrongContext);
   const { setIsAppLoading } = useContext(AppLoadingContext);
 
   const {
@@ -73,7 +71,7 @@ export default function CommunicationLog({ regionId, recipientId }) {
 
         setLogs(response);
       } catch (err) {
-        setErrorResponseCode(err.status);
+        setError('Error fetching communication logs');
       } finally {
         setIsAppLoading(false);
       }
